+++ Runtime/Events/EventBusManager.cs
@@ -1,12 +1,9 @@
 using System;
+using System.Linq;
 using System.Reflection;
 using System.Collections;
 using System.Collections.Generic;
 using UnityEngine;
-
-#if UNITY_EDITOR
-using UnityEditor;
-#endif
 
 using CinderUtils.Extensions;
 using CinderUtils.Reflection;
@@ -15,56 +12,39 @@ using CinderUtils.Reflection;
 namespace CinderUtils.Events {
 
     public static class EventBusManager {
-        internal static HashSet<Type> eventTypes;
-        internal static HashSet<Type> eventBusTypes;
+        internal static HashSet<Type> eventTypes = new();
+        internal static HashSet<Type> eventBusTypes = new();
 
         public static IReadOnlyCollection<Type> EventTypes { get => eventTypes; }
 
         [RuntimeInitializeOnLoadMethod(RuntimeInitializeLoadType.AfterAssembliesLoaded)]
         internal static void Initialize() {
-            eventTypes = AssemblyUtils.GetSubtypesOf<IEvent>(false);
-            eventBusTypes = GetEventBusTypes();
+            GetEventTypes();
+            GetEventBusTypes();
 
-            CinderDebug.Log("CinderUtils.Events: EventBusManager initialized.");
+            CinderDebug.Log("CinderUtils: EventBus: Initialized.");
         }
 
-#if UNITY_EDITOR
-        static PlayModeStateChange PlayModeState { get; set; }
-
-        [InitializeOnLoadMethod]
-        static void InitializeEditor() {
-            EditorApplication.playModeStateChanged -= OnPlayModeStateChanged;
-            EditorApplication.playModeStateChanged += OnPlayModeStateChanged;
+        static void GetEventTypes() { 
+            AssemblyUtils.GetSubtypesOf<IEvent>().ForEach(t => eventTypes.Add(t));
         }
 
-        static void OnPlayModeStateChanged(PlayModeStateChange state) {
-            PlayModeState = state;
-            if (state == PlayModeStateChange.ExitingPlayMode) {
-                ClearBuses();
-            }
-        }
-#endif
-
-
-        static HashSet<Type> GetEventBusTypes() {
-            HashSet<Type> eventBusTypes = new();
-
+        static void GetEventBusTypes() {
             var baseBusType = typeof(EventBus<>);
-            foreach (var eventType in eventTypes) {
+
+            foreach (var eventType in EventTypes) {
                 var busType = baseBusType.MakeGenericType(eventType);
                 eventBusTypes.Add(busType);
             }
-
-            return eventBusTypes;
         }
 
-        static void ClearBuses() {
+        public static void ClearBuses() {
             foreach (var busType in eventBusTypes) {
                 var clearMethod = busType.GetMethod("Clear", BindingFlags.Static | BindingFlags.NonPublic);
                 clearMethod?.Invoke(null, null);
             }
 
-            CinderDebug.Log("CinderUtils.Events: EventBusManager: Event buses cleaned.");
+            CinderDebug.Log("CinderUtils: EventBus: Buses reset.");
         }
     }