+++ Runtime/Extensions/TypeExtensions.cs
@@ -1,5 +1,6 @@
 using System;
 using System.Reflection;
+using UnityEngine;
 
 
 namespace CinderUtils.Extensions {
@@ -9,8 +10,18 @@ namespace CinderUtils.Extensions {
 
         public static bool Is<T>(this Type type) {
             Type baseType = typeof(T);
+            
+            // Could a variable of type "parentType" get assigned a value of the type "type"?
+            // If so, the type is a subtype
             return baseType.IsAssignableFrom(type);
         }
-    }
 
+        public static bool HasAttribute<TAttr>(this Type type) where TAttr : Attribute {
+            return type.GetCustomAttribute<TAttr>() != null;
+        }
+
+        public static bool IsMonoBehaviour(this Type t) {
+            return t.Is<MonoBehaviour>();
+        }
+    }
 }